backend/snippets/serializers.py
from rest_framework import serializers
from backend.snippets.models import Snippet


class SnippetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Snippet
        fields = ['id', 'title', 'code', 'language', 'description', "category"]
